requirements.txt
fastapi
uvicorn
httpx
pandas
openpyxl
python-dotenv
jinja2
python-multipart
sqlalchemy
mysqlclient
